requirements.txt
fastapi
uvicorn
langchain
# langchain-openai
langchain-ollama
python-dotenv
pydantic
openai
streamlit
ollama